# telemetry/collector/src/main/java/ru/practicum/collector/model/hub/ScenarioAddedEvent.java
package ru.practicum.collector.model.hub;

import jakarta.validation.Valid;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotEmpty;
import jakarta.validation.constraints.Size;
import lombok.AccessLevel;
import lombok.Getter;
import lombok.Setter;
import lombok.ToString;
import lombok.experimental.FieldDefaults;

import java.util.List;

@Getter
@Setter
@ToString(callSuper = true)
@FieldDefaults(level = AccessLevel.PRIVATE)
public class ScenarioAddedEvent extends HubEvent {
    @NotBlank
    @Size(min = 3, message = "name must contain at least 3 characters")
    String name;

    @NotEmpty
    @Valid
    List<ScenarioCondition> conditions;

    @NotEmpty
    @Valid
    List<DeviceAction> actions;
}